services/ingestor_news/requirements.txt
feedparser==6.0.*
httpx==0.27.*
asyncpg==0.29.*
redis==5.0.*
pyyaml==6.0.*
